creating a solution and a project.
