+++ src/Asteroids.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Asteroids", "Asteroids\Asteroids.csproj", "{CFEA5AAE-76C5-466F-8435-44F722BC7093}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CFEA5AAE-76C5-466F-8435-44F722BC7093}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CFEA5AAE-76C5-466F-8435-44F722BC7093}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CFEA5AAE-76C5-466F-8435-44F722BC7093}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CFEA5AAE-76C5-466F-8435-44F722BC7093}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
