+++ src/Asteroids/score
@@ -1,0 +1,1 @@
+55976948H2ACBB4C1H2ACBB487
